Исправления после 3-го Pull-request.
