+++ src/main/java/ru/egor/tgBot/controllers/ProductRestController.java
@@ -1,11 +1,12 @@
 package ru.egor.tgBot.controllers;
 
-import org.apache.coyote.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ru.egor.tgBot.entity.Product;
+import ru.egor.tgBot.services.OrderProductService;
 import ru.egor.tgBot.services.ProductService;
 
 import java.util.List;
@@ -16,15 +17,27 @@ public class ProductRestController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private OrderProductService orderProductService;
+
     @GetMapping("/rest/products/search")
     public List<Product> searchProduct(@RequestParam(required = false) Long categoryId,
                                        @RequestParam(required = false) String name) {
         if (name != null) {
             return productService.findByNameContainingIgnoreCase(name);
-        } else if (categoryId != null) {
+        } else
             return productService.getProductsByCategoryName(categoryId);
-        }
-        return null;
+
+    }
+
+    @GetMapping("/rest/clients/{id}/products")
+    public List<Product> findAllProductsByClientId(@PathVariable Long id) {
+        return orderProductService.findAllProductsByClientId(id);
+    }
+
+    @GetMapping("/rest/products/popular")
+    public List<Product> getTopPopularProducts(@RequestParam int limit){
+        return orderProductService.getTopPopularProducts(limit);
     }
 
 }
