+++ src/main/java/ru/egor/tgBot/entity/Category.java
@@ -3,8 +3,6 @@ package ru.egor.tgBot.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Data
