+++ src/main/java/ru/egor/tgBot/services/OrderProductService.java
@@ -1,14 +1,11 @@
 package ru.egor.tgBot.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.egor.tgBot.entity.Product;
 import ru.egor.tgBot.repository.OrderProductRepository;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class OrderProductService {
@@ -21,10 +18,6 @@ public class OrderProductService {
     }
 
     public List<Product> getTopPopularProducts(int limit) {
-        Pageable pageable = PageRequest.of(0, limit);
-        return orderProductRepository.findTopPopularProducts(pageable)
-                .stream()
-                .map(arr -> (Product) arr[0])
-                .collect(Collectors.toList());
+        return orderProductRepository.findTopPopularProducts(limit);
     }
 }
